+++ CHANGES.md
@@ -1,3 +1,8 @@
+# 5.0 (December 2018)
+- Upgrade to Sentio 5.0 - See Sentio's change log for details, includes update to d3 v5
+- Upgrade to Angular.io v7 - Includes updates to Rxjs and Angular
+- Upgrade to D3 v5
+
 # 3.0 (January 2017)
 
 ## Sentio 3.0